[DEV-7216] Rename new file sharedComponents/Page to PageWrapper and update imports
