+++ src/js/components/account/Account.jsx
@@ -10,7 +10,7 @@ import { ShareIcon } from 'data-transparency-ui';
 import * as MetaTagHelper from 'helpers/metaTagHelper';
 import { getBaseUrl, handleShareOptionClick } from 'helpers/socialShare';
 
-import PageWrapper from 'components/sharedComponents/Page';
+import PageWrapper from 'components/sharedComponents/PageWrapper';
 
 import AccountOverview from './AccountOverview';
 import SearchSidebar from './SearchSidebar';
