+++ src/js/components/agencyLanding/AgencyLandingPage.jsx
@@ -9,7 +9,7 @@ import { ShareIcon } from 'data-transparency-ui';
 import { agencyLandingPageMetaTags } from 'helpers/metaTagHelper';
 import { getBaseUrl, handleShareOptionClick } from 'helpers/socialShare';
 
-import PageWrapper from 'components/sharedComponents/Page';
+import PageWrapper from 'components/sharedComponents/PageWrapper';
 
 import AgencyLandingContainer from 'containers/agencyLanding/AgencyLandingContainer';
